refactor: create clihelpers/ package with shared CLI helpers
Extract truly shared CLI helpers from cli/common.py into clihelpers/:
- clihelpers/output.py — OutputFormat enum
- clihelpers/config.py — load_config_or_exit
- clihelpers/endpoints.py — resolve_endpoints, build_endpoint_filter
- clihelpers/progress.py — VolumeProgressBar, CheckProgressBar

cli/common.py becomes a transitional shim that re-exports from
clihelpers and keeps domain-specific helpers (managed_mount,
check_all_with_progress, etc.) until they move to domain CLIs.
