+++ nbkp/clihelpers/endpoints.py
@@ -1,0 +1,71 @@
+"""Endpoint resolution helpers for CLI commands."""
+
+from __future__ import annotations
+
+import typer
+
+from ..config import Config
+from ..config.epresolution import (
+    EndpointFilter,
+    NetworkType,
+    ResolvedEndpoints,
+)
+from ..remote.resolution import resolve_all_endpoints
+
+
+def build_endpoint_filter(
+    locations: list[str] | None,
+    exclude_locations: list[str] | None,
+    network: NetworkType | None,
+) -> EndpointFilter | None:
+    """Build an EndpointFilter from CLI options."""
+    locs = locations or []
+    excl = exclude_locations or []
+    return (
+        EndpointFilter(locations=locs, exclude_locations=excl, network=network)
+        if locs or excl or network is not None
+        else None
+    )
+
+
+def _validate_locations(
+    cfg: Config,
+    locations: list[str] | None,
+    exclude_locations: list[str] | None,
+) -> None:
+    """Exit with an error if any location value is not defined in the config."""
+    known = set(cfg.known_locations())
+    if not known:
+        all_values = [*(locations or []), *(exclude_locations or [])]
+        if all_values:
+            typer.echo(
+                "Error: no locations are defined in the configuration."
+                " --location and --exclude-location cannot be used.",
+                err=True,
+            )
+            raise typer.Exit(2)
+        return
+    for label, values in [
+        ("--location", locations),
+        ("--exclude-location", exclude_locations),
+    ]:
+        for v in values or []:
+            if v not in known:
+                typer.echo(
+                    f"Error: unknown location '{v}' passed to {label}."
+                    f" Known locations: {', '.join(sorted(known))}",
+                    err=True,
+                )
+                raise typer.Exit(2)
+
+
+def resolve_endpoints(
+    cfg: Config,
+    locations: list[str] | None,
+    exclude_locations: list[str] | None,
+    network: NetworkType | None,
+) -> ResolvedEndpoints:
+    """Build filter and resolve all endpoints once."""
+    _validate_locations(cfg, locations, exclude_locations)
+    ef = build_endpoint_filter(locations, exclude_locations, network)
+    return resolve_all_endpoints(cfg, ef)
