+++ nbkp/clihelpers/output.py
@@ -1,0 +1,12 @@
+"""Output format enum shared across CLI commands."""
+
+from __future__ import annotations
+
+import enum
+
+
+class OutputFormat(str, enum.Enum):
+    """Output format for CLI commands."""
+
+    HUMAN = "human"
+    JSON = "json"
